pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lexdrift"
version = "0.1.0"
description = "SEC Semantic Drift Analyzer — track how public companies change their language across filings"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    # API
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    # HTTP client
    "httpx>=0.27.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.30",
    "aiosqlite>=0.20.0",
    "alembic>=1.13.0",
    # HTML parsing
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    # NLP
    "spacy>=3.7.0",
    "sentence-transformers>=3.0.0",
    # Sentiment lexicon & data
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "scipy>=1.13.0",
    # Task queue
    "celery[redis]>=5.4.0",
    # Config
    "pydantic-settings>=2.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx",  # for TestClient
    "ruff>=0.5.0",
]
prod = [
    "asyncpg>=0.29.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/lexdrift"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 100
